requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pygit2
dulwich
gunicorn
python-dotenv
whitenoise